Make withSecrets parallel-safe
Our secret management now works by cleaning up on exit, rather than mucking
about with environ.  This is nice but it means that having multiple secrets
blocks in parallel could work incorrectly, because the first to exit would
clean up secrets, even if others were still using them!  So now we track how
many blocks are active, and only clean up on exiting the last one.

I also updated the docs, which I had forgotten to do for the previous changes.

Test plan: deploy using this

Auditors: csilvers

+++ vars/withSecrets.groovy
@@ -2,9 +2,15 @@
 
 // This is no longer where we store secrets.py, it's just where the
 // password lives.
-def _secretsPasswordDir() {
-   return "${env.HOME}/secrets_py";
-}
+_SECRETS_PASSWORD_DIR = "${env.HOME}/secrets_py";
+
+// The number of active withSecrets blocks.  We only want to clean secrets up
+// at the end if we are exiting the last withSecrets block, since they likely
+// all share a workspace.
+// TODO(benkraft): In principle this should be per-directory; in practice
+// assuming we're always in the same directory is good enough at present.
+// TODO(benkraft): Make sure updates to this are actually atomic.
+_activeSecretsBlocks = 0;
 
 
 // This must be called from workspace-root.
@@ -19,25 +25,32 @@ def call(Closure body) {
       } else {
          webappSecretsDir = "webapp";
       }
+
       // First, set up secrets.
       // This decryption command was modified from the make target
       // "secrets_decrypt" in the webapp project.
+      // Note that we do this even if ACTIVE_SECRETS_BLOCKS > 0;
+      // secrets.py.cast5 might have changed.
       exec(["openssl", "cast5-cbc", "-d",
             "-in", "${webappSecretsDir}/secrets.py.cast5",
             "-out", "${webappSecretsDir}/secrets.py",
-            "-kfile", "${_secretsPasswordDir()}/secrets.py.cast5.password"]);
+            "-kfile", "${_SECRETS_PASSWORD_DIR}/secrets.py.cast5.password"]);
       sh("chmod 600 ${webappSecretsDir}/secrets.py");
+      _activeSecretsBlocks++;
 
       // Then, tell alertlib where secrets live, and run the wrapped block.
       withEnv(["ALERTLIB_SECRETS_DIR=${webappSecretsDir}"]){
          body();
       }
    } finally {
-      // Finally, clean up secrets.py so if the next job intends
-      // to run without secrets, it does.  We remove both versions
-      // just in case an old one was floating around.
-      sh("rm -f webapp/shared/secrets.py");
-      sh("rm -f webapp/secrets.py");
+      _activeSecretsBlocks--;
+      // Finally, iff we're exiting the last withSecrets block, clean up
+      // secrets.py so if the next job intends to run without secrets, it does.
+      // We remove both versions just in case an old one was floating around.
+      if (!activeSecretsBlocks) {
+         sh("rm -f webapp/shared/secrets.py");
+         sh("rm -f webapp/secrets.py");
+      }
    }
 }
